refactor: FakeFileProvider, ApiDataFixture 폴더 이동
