+++ android/Staccato_AN/app/src/test/java/com/on/staccato/data/ApiResultCallAdapterTest.kt
@@ -2,16 +2,6 @@ package com.on.staccato.data
 
 import com.on.staccato.CoroutinesTestExtension
 import com.on.staccato.StaccatoApplication.Companion.retrofit
-import com.on.staccato.createErrorBy400
-import com.on.staccato.createErrorBy401
-import com.on.staccato.createErrorBy403
-import com.on.staccato.createErrorBy413
-import com.on.staccato.createErrorBy500
-import com.on.staccato.createFakeImageFile
-import com.on.staccato.createInvalidMemoryRequest
-import com.on.staccato.createMemoryCreationResponse
-import com.on.staccato.createMemoryResponse
-import com.on.staccato.createValidMemoryRequest
 import com.on.staccato.data.comment.CommentApiService
 import com.on.staccato.data.dto.image.ImageResponse
 import com.on.staccato.data.dto.memory.MemoryCreationResponse
